Proxy / Combine proxy.
Once secured, use jetty:run -Dgeonetwork.ESFeaturesProxy.username=elastic -Dgeonetwork.ESFeaturesProxy.password=elastic -Des.username=elastic -Des.password=elastic to set authentication.
