+++ web/src/main/webResources/WEB-INF/web.xml
@@ -406,7 +406,7 @@
   -->
   <servlet>
     <servlet-name>ESFeaturesProxy</servlet-name>
-    <servlet-class>org.mitre.dsmiley.httpproxy.URITemplateProxyServlet</servlet-class>
+    <servlet-class>org.fao.geonet.proxy.URITemplateProxyServlet</servlet-class>
     <init-param>
       <param-name>targetUri</param-name>
       <param-value>${es.url}/${es.index.features}/{_}</param-value>
@@ -415,19 +415,11 @@
       <param-name>log</param-name>
       <param-value>false</param-value>
     </init-param>
-  </servlet>
-  <!--
-  If Elasticsearch has a security layer, create a readonly user
-  with access to the features index and configure the proxy
-  with basic authentication if needed. Example:
-  <servlet>
-    <servlet-name>ESFeaturesProxy</servlet-name>
-    <servlet-class>org.fao.geonet.proxy.EsBasicAuthInjectorProxyServlet</servlet-class>
-    <init-param>
-      <param-name>targetUri</param-name>
-      <param-value>${es.url}/${es.index.features}/{_}</param-value>
-    </init-param>
-    <init-param>
+    <!--
+     If Elasticsearch has a security layer, create a readonly user
+     with access to the features index and configure the proxy
+     with basic authentication if needed.-->
+    <!--<init-param>
       <param-name>username</param-name>
       <param-value>${es.ro.username}</param-value>
     </init-param>
@@ -436,20 +428,15 @@
       <param-value>${es.ro.password}</param-value>
     </init-param>
     <init-param>
-      <param-name>log</param-name>
-      <param-value>false</param-value>
-    </init-param>
-    <init-param>
       <param-name>isSecured</param-name>
       <param-value>true</param-value>
-    </init-param>
+    </init-param>-->
   </servlet>
-  -->
-
   <servlet-mapping>
     <servlet-name>ESFeaturesProxy</servlet-name>
     <url-pattern>/index/features</url-pattern>
   </servlet-mapping>
+
 
   <servlet>
     <servlet-name>HttpDashboardProxy</servlet-name>
